+++ src/routes/TraderForm/Form/NotionalSwap/NotionalSwap.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import React, { useState } from 'react';
 
 import { NotionalSwap } from './index';
 
@@ -10,13 +9,7 @@ export default {
 } as ComponentMeta<typeof NotionalSwap>;
 
 const Template: ComponentStory<typeof NotionalSwap> = (args) => {
-  const [mode, setMode] = useState<'fixed' | 'variable'>('fixed');
-  return <NotionalSwap {...args} mode={mode} onSwap={setMode} />;
+  return <NotionalSwap />;
 };
 
 export const Default = Template.bind({});
-Default.args = {
-  fixedRate: 5.49,
-  variableRate: 2.49,
-  mode: 'fixed',
-};
