+++ src/routes/TraderForm/index.tsx
@@ -1,14 +1,7 @@
 import { Page as BrokoliPage } from 'brokoli-ui';
 import * as React from 'react';
-import { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 
-import {
-  getAvailableNotionalsThunk,
-  getFixedRateThunk,
-  getVariableRateThunk,
-} from '../../app/features/swap-form/thunks';
-import { useAppDispatch } from '../../app/hooks';
 import { Loading } from '../../components/atomic/Loading/Loading';
 import { Panel } from '../../components/atomic/Panel/Panel';
 import { useSwapFormAMM } from '../../hooks/useSwapFormAMM';
@@ -27,18 +20,8 @@ import {
 } from './TraderForm.styled';
 
 export const TraderFormRoute: React.FunctionComponent = () => {
-  const dispatch = useAppDispatch();
   const { form } = useParams();
   const { aMM, loading, error, idle } = useSwapFormAMM();
-
-  useEffect(() => {
-    if (!aMM) {
-      return;
-    }
-    void dispatch(getFixedRateThunk());
-    void dispatch(getVariableRateThunk());
-    void dispatch(getAvailableNotionalsThunk());
-  }, [dispatch, aMM]);
 
   let pageContent = (
     <MainAndFormSectionBox data-testid="BrokoliPage-MainAndFormSectionBox">
